// src/types/product.ts
// Generated by https://quicktype.io

import { ReactNode } from "react";

export interface Product {
    id:          number;
    title:       string;
    price:       number;
    description: string;
    category:    string;
    image:       string;
    rating:      Rating;
}

export interface Rating {
    rate:  number;
    count: number;
}


export type ProductContextType = {
    products: Product[],
    product: Product | null,
    isLoading?: boolean,
    loadProducts: () => void
    deleteProduct: (product: Product) => void
    
}

export type ProductProviderProps = {
    children: ReactNode
}

export type ProductActions = 
    | { type: 'LOAD_PRODUCTS', payload: Product[] }
    | { type: 'GET_PRODUCTS', payload?: Product }
    | { type: 'DELETE_PRODUCT', payload: Product }
    | { type: 'LOADING', payload: boolean }

export type ProductState = {
    products: Product[],
    product: Product | null,
    isLoading: boolean
}